*** 2007-12-03 by Roland Schulz - /settrait command now takes target argument like most other GM commands and   help cleanups, thanks to Vornne - Fixed some warnings.

+++ trunk/src/common/util/skillcache.cpp
@@ -288,7 +288,7 @@ psSkillCacheItem *psSkillCache::getItemBySkillId(uint id)
     while (p.HasNext())
     {
         psSkillCacheItem *item = p.Next();
-        if (item && item->getSkillId() == id)
+        if (item && (uint)item->getSkillId() == id)
             return item;
     }
     return NULL;
